backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
cohere>=4.0
mcp>=0.1.0
better-auth>=0.0.1b11
sqlmodel>=0.0.8
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
asyncpg>=0.25.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.23.0